Break SiteMembershipCreator off of SiteCreator
The SiteCreator service was doing double duty, creating sites and adding
users to sites. This commit creates a new service and moves the code to
add a user to a site into that service.

One change in the API as a result of this is that now attempting to
create a site that already exists renders an error instead of adding a
user to the site. The addUser endpoint will need to be used now to add
users to a site.